processing-worker/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx>=0.24.0,<0.25.0

# Environment and configuration
python-dotenv==1.0.0

# Google Cloud services  
google-cloud-pubsub==2.18.4

# Database/Storage
supabase==2.0.2

# AI/ML
openai==1.3.7

# Data processing
pydantic==2.5.0
typing-extensions==4.8.0

# Logging and utilities
structlog==23.2.0

# JSON and data validation
jsonschema==4.20.0

# Date and time utilities
python-dateutil==2.8.2

# Additional utilities
requests==2.31.0